+++ src/main/java/akin/city_card/wallet/repository/WalletTransferRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface WalletTransferRepository extends JpaRepository<WalletTransfer, Integer> {
+public interface WalletTransferRepository extends JpaRepository<WalletTransfer, Long> {
 }
